Develope

+++ app/Http/Controllers/NovelGroupController.php
@@ -2,11 +2,14 @@
 
 namespace App\Http\Controllers;
 
+use App\Events\NewSpeedEvent;
 use App\Favorite;
 use App\Keyword;
 use App\Mailbox;
 use App\MailLog;
 use App\Novel;
+use App\NovelGroupHashTag;
+use App\NovelGroupKeyword;
 use App\User;
 use Auth;
 use Carbon\Carbon;
@@ -39,10 +42,20 @@ class NovelGroupController extends Controller
     {
         if (Auth::user()->isAdmin()) {
             //if you are admin
-            $novel_groups = NovelGroup::with('novels')->latest()->get();
+            $novel_groups = NovelGroup::with('novels')->latest()->paginate(10);
+
+            if ($request->url('/admin/recommendations')) {
+                $novel_groups = NovelGroup::with('novels')->orderBy('recommend_order', 'asc')->latest()->paginate(10);
+
+                //make recommend_order to null when recommend_order > 5 [max valid order is 5]
+                foreach( $novel_groups as  $novel_group){
+                    if($novel_group->recommend_order > 5) $novel_group->recommend_order = null;
+                }
+            }
+
         } else {
             //if you are user
-            $novel_groups = $request->user()->novel_groups()->with('novels')->latest()->get();
+            $novel_groups = $request->user()->novel_groups()->with('novels')->latest()->paginate(10);
         }
 
         //check an agreement agreed or not
@@ -71,10 +84,9 @@ class NovelGroupController extends Controller
                 foreach ($novel->comments as $commenat) {
                     $comments_count++;
                 }
-                foreach ($novel->reviews as $n) {
-                    $review_count++;
-                }
-
+            }
+            foreach ($novel_group->reviews as $n) {
+                $review_count++;
             }
             // 소설이 없다면
             if ($novel_group->novels->count() != 0) {
@@ -97,8 +109,7 @@ class NovelGroupController extends Controller
         if (!isset($request->page)) {
             $request->page = 1;
         }
-        $novel_group_per_page = $novel_groups->forPage($request->page, 5);
-        $novel_groups = new LengthAwarePaginator($novel_group_per_page, $novel_groups_count, 5);
+
 
         return \Response::json(['novel_groups' => $novel_groups, 'count_data' => $count_data, 'review_count_data' => $review_count_data, 'latested_at' => $latested_at, 'author' => $author]);
         // dd($user_novels);
@@ -122,28 +133,73 @@ class NovelGroupController extends Controller
      */
     public function store(Request $request)
     {
-        //
+
+        //   dd($request->all());
+
+
         Validator::make($request->all(), [
-            'nickname' => 'required|max:255',
-            'title' => 'required',
+            'nickname_id' => 'required',
+            'title' => 'required|max:255',
             'description' => 'required',
+            'hash_tags' => 'required',
+            /* 'keyword2' => 'required',
+             'keyword3' => 'required',
+             'keyword4' => 'required',
+             'keyword5' => 'required',
+             'keyword6' => 'required',
+             'keyword7' => 'required',*/
             'cover_photo' => 'mimes:jpeg,png|image|max:1024|dimensions:max_width=1080,max_height=1620',
             'cover_photo2' => 'mimes:jpeg,png|image|max:1024|dimensions:max_width=1080,max_height=1080',
 
         ], [
-            'nickname.required' => '필명은 필수 입니다.',
+            'nickname_id.required' => '필명은 필수 입니다.',
             'title.required' => '제목은 필수 입니다.',
+            'title.max' => '제목이 너무 깁니다.',
             'description.required' => '설명은 필수 입니다.',
             'cover_photo.dimensions' => '표지1 크기는 1080*1620 이어야 합니다',
             'cover_photo.max' => '표지1 용량은 1M를 넘지 않아야 합니다',
             'cover_photo2.dimensions' => '표지2 크기는 1080*1080 이어야 합니다',
             'cover_photo2.max' => '표지2 용량은 1M를 넘지 않아야 합니다',
+            'hash_tags.required' => '첫번째 키워드를 입력해 주세요',
+            /*  'keyword2.required' => '두번째 키워드를 입력해 주세요',
+              'keyword3.required' => '세번째 키워드를 입력해 주세요',
+              'keyword4.required' => '네번째 키워드를 입력해 주세요',
+              'keyword5.required' => '다섯번째 키워드를 입력해 주세요',
+              'keyword6.required' => '여섯번째 키워드를 입력해 주세요',
+              'keyword7.required' => '일곱번째 키워드를 입력해 주세요',*/
 
         ])->validate();
 
         $input = $request->all();
         //if validation is passed then insert the record
-        $new_novel_group = $request->user()->novel_groups()->create($input);
+//        $new_novel_group = $request->user()->novel_groups()->create($input);
+
+        $new_novel_group = new NovelGroup();
+        $new_novel_group->nickname_id = $request->nickname_id;
+        $new_novel_group->user_id = $request->user()->id;
+        $new_novel_group->title = $request->title;
+        $new_novel_group->description = $request->description;
+        $new_novel_group->cover_photo = $request->cover_photo;
+        $new_novel_group->cover_photo2 = $request->cover_photo2;
+        $new_novel_group->save();
+
+
+        //save the type
+        $new_novel_group_keyword = new NovelGroupKeyword();
+        $new_novel_group_keyword->novel_group_id = $new_novel_group->id;
+        $new_novel_group_keyword->keyword_id = $request->input('keyword1');
+        $new_novel_group_keyword->save();
+
+        //save the has tags
+        //  for ($i = 2; $i <= 7; $i++) {
+        foreach ($input['hash_tags'] as $hash_tag) {
+            $new_novel_group_hash_tag = new NovelGroupHashTag();
+            $new_novel_group_hash_tag->novel_group_id = $new_novel_group->id;
+            $new_novel_group_hash_tag->tag = $hash_tag;
+            $new_novel_group_hash_tag->save();
+        }
+        //  }
+
 
         //upload the picture
         if ($request->hasFile('cover_photo') or $request->hasFile('cover_photo2')) {
@@ -184,12 +240,16 @@ class NovelGroupController extends Controller
             $new_novel_group->save();
         }
 
+
+        flash("생성을 성공했습니다");
+        //  return redirect()->route('author_novel_group', ['id' => $new_novel_group->id]);
+
+        event(new NewSpeedEvent("new_novel_group", "작가 " . $new_novel_group->nicknames->nickname . "의 신작 " . $new_novel_group->title . " 이(가) 신규 등록 되었습니다.", route('each_novel.novel_group', ['id' => $new_novel_group->id]), "/img/novel_covers/" . $new_novel_group->cover_photo, $new_novel_group->id));
+
         if ($request->ajax()) {
             return "OK";
         }
 
-        flash("생성을 성공했습니다");
-        //  return redirect()->route('author_novel_group', ['id' => $new_novel_group->id]);
         return redirect()->route('author_index');
     }
 
@@ -224,27 +284,32 @@ class NovelGroupController extends Controller
     {
 
         // $novel_group= $request->user()->novel_groups()->with('users.nicknames')->where('id',$id)->first();
-        $novel_group = NovelGroup::where('id', $id)->first();
+        $novel_group = NovelGroup::where('id', $id)->with('nicknames', 'keywords', 'hash_tags')->first();
         $nicknames = $request->user()->nicknames()->get();
+
+        $selected_hash_tags = NovelGroupHashTag::where('novel_group_id', $id)->pluck('tag');
         $keyword1 = Keyword::select('id', 'name')->where('category', '1')->get();
-        $keyword2 = Keyword::select('id', 'name')->where('category', '2')->get();
-        $keyword3 = Keyword::select('id', 'name')->where('category', '3')->get();
-        $keyword4 = Keyword::select('id', 'name')->where('category', '4')->get();
-        $keyword5 = Keyword::select('id', 'name')->where('category', '5')->get();
-        $keyword6 = Keyword::select('id', 'name')->where('category', '6')->get();
-        $keyword7 = Keyword::select('id', 'name')->where('category', '7')->get();
+        $hash_tag_keywords = Keyword::select('id', 'name')->where('category', '<>', '1')->get();
+        /* $keyword3 = Keyword::select('id', 'name')->where('category', '3')->get();
+         $keyword4 = Keyword::select('id', 'name')->where('category', '4')->get();
+         $keyword5 = Keyword::select('id', 'name')->where('category', '5')->get();
+         $keyword6 = Keyword::select('id', 'name')->where('category', '6')->get();
+         $keyword7 = Keyword::select('id', 'name')->where('category', '7')->get();*/
 
 
         return \Response::json([
             'novel_group' => $novel_group,
             'nick_names' => $nicknames,
+            'selected_hash_tags' => $selected_hash_tags,
             'keyword1' => $keyword1,
-            'keyword2' => $keyword2, 'keyword3' => $keyword3,
-            'keyword4' => $keyword4, 'keyword5' => $keyword5,
-            'keyword6' => $keyword6, 'keyword7' => $keyword7,
+            'hash_tag_keywords' => $hash_tag_keywords,
+            /* 'keyword3' => $keyword3,
+              'keyword4' => $keyword4, 'keyword5' => $keyword5,
+              'keyword6' => $keyword6, 'keyword7' => $keyword7,*/
 
 
         ]);
+        //return redirect()->route('author.edit',compact('novel_group','nicknames','selected_hash_tags','keyword1', 'hash_tag_keywords'));
     }
 
     /**
@@ -257,19 +322,22 @@ class NovelGroupController extends Controller
     public function update(Request $request, $id)
     {
         //
-
-        $input = $request->except('_token', '_method', 'default_cover_photo');
+        //$input = $request->except('_token', '_method', 'default_cover_photo');
+        $input = $request->only('nickname_id', 'title', 'description', 'cover_photo', 'cover_photo2');
+        $hash_tags = $request->only('hash_tags');
+        $keywords = $request->only('keyword1');
 
         Validator::make($request->all(), [
-            'nickname' => 'required|max:255',
-            'title' => 'required',
+            'nickname_id' => 'required',
+            'title' => 'required|max:255',
             'description' => 'required',
             'cover_photo' => 'mimes:jpeg,png|image|max:1024|dimensions:max_width=1080,max_height=1620',
             'cover_photo2' => 'mimes:jpeg,png|image|max:1024|dimensions:max_width=1080,max_height=1080',
 
         ], [
-            'nickname.required' => '필명은 필수 입니다.',
+            'nickname_id.required' => '필명은 필수 입니다.',
             'title.required' => '제목은 필수 입니다.',
+            'title.max' => '제목이 너무 깁니다.',
             'description.required' => '설명은 필수 입니다.',
             'cover_photo.dimensions' => '표지1 크기는 1080*1620 이어야 합니다',
             'cover_photo.max' => '표지1 용량은 1M를 넘지 않아야 합니다',
@@ -304,15 +372,60 @@ class NovelGroupController extends Controller
             }
 
         } else if ($request->default_cover_photo) {
-
             $input['cover_photo'] = "default_" . $request->default_cover_photo . ".jpg";
-        } /*else {
+        } else {
             // $new_novel_group = $request->user()->novel_groups()->create($input);
-            $input['cover_photo'] = "default_.jpg";
-        }*/
+            $novel_group = NovelGroup::find($id);
 
+            if ($novel_group->cover_photo == null) {
+
+                $input['cover_photo'] = "default_.jpg";
+            } else {
+                $input['cover_photo'] = $novel_group->cover_photo;
+            }
+        }
+
+        //update the novel_group
         NovelGroup::where('id', $id)->update($input);
-        //redirect to novels
+
+        //update the novel_group type
+        NovelGroupKeyword::where('novel_group_id', $id)->update(['keyword_id' => $keywords['keyword1']]);
+
+        //Delete the hash tags which are removed by user
+        // get all group hash tags
+        $selected_hash_tags = NovelGroupHashTag::where('novel_group_id', $id)->get();
+        $deleteFlag = false;
+        //if a selected hash tag is not in new group hash tags [updated hash tags] then delete that from db
+        foreach ($selected_hash_tags as $selected_hash_tag) {
+            foreach ($hash_tags['hash_tags'] as $hash_tag) {
+                if ($selected_hash_tag->tag != $hash_tag) {
+                    $deleteFlag = true;
+
+                } else {
+                    $deleteFlag = false;
+                    break;
+                }
+            }
+
+            if ($deleteFlag) {
+                NovelGroupHashTag::where('id', $selected_hash_tag->id)->delete();
+            }
+
+        }
+
+//Update and Insert new Tags
+        foreach ($hash_tags['hash_tags'] as $hash_tag) {
+            //Check if already exists or not
+            $already_exists = NovelGroupHashTag::where(['novel_group_id' => $id, 'tag' => $hash_tag])->first();
+            //if don't exist then insert it
+            if (!$already_exists) {
+                NovelGroupHashTag::create(['novel_group_id' => $id, 'tag' => $hash_tag]);
+            }
+
+        }
+
+
+//redirect to novels
         flash("수정을 성공했습니다");
         if ($request->ajax()) {
             return \Response::json(['status' => 'ok']);
@@ -327,14 +440,22 @@ class NovelGroupController extends Controller
      * @param  int $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public
+    function destroy($id)
     {
-        //
-        $novel_group = NovelGroup::find($id);
-        $novel_group->delete();
+        $novel = Novel::where('novel_group_id', $id)->first();
+        if ($novel == null) {
+            $novel_group = NovelGroup::find($id);
+            $novel_group->delete();
+        } else {
+            return response()->json(['error' => 1, 'message' => '해당 소설을 삭제할 수 없습니다. 각 회차를 먼저 삭제해 주십시오.']);
+        }
+
+        return response()->json(['error' => 0, 'message' => '삭제 되었습니다.']);
     }
 
-    public function secret($id)
+    public
+    function secret($id)
     {
         $novel_group = NovelGroup::findOrFail($id);
 
@@ -349,8 +470,8 @@ class NovelGroupController extends Controller
 
 
         $new_mail = new Mailbox();
-        $new_mail->subject = "'" . $novel_group->title . "'이 비밀이 됩니다.";
-        $new_mail->body = $novel_group->secret . " '" . $novel_group->title . "'이 비밀이 됩니다.";
+        $new_mail->subject = "'" . $novel_group->title . "'이(가) 비밀 글이 됩니다.";
+        $new_mail->body = $novel_group->secret . " 이후로 '" . $novel_group->title . "'이(가) 비밀 글이 됩니다.";
         $new_mail->from = Auth::user()->id;
         $new_mail->novel_group_id = $novel_group->id;
         $new_mail->save();
@@ -368,7 +489,8 @@ class NovelGroupController extends Controller
 
     }
 
-    public function non_secret($id)
+    public
+    function non_secret($id)
     {
 
         $novel_group = NovelGroup::findOrFail($id);
@@ -380,8 +502,8 @@ class NovelGroupController extends Controller
 
 
         $new_mail = new Mailbox();
-        $new_mail->subject = "'" . $novel_group->title . "'이 비밀 해제 됩니다.";
-        $new_mail->body = $novel_group->secret . " '" . $novel_group->title . "'이 비밀 해제 됩니다.";
+        $new_mail->subject = "'" . $novel_group->title . "'이(가) 비밀 해제 됩니다.";
+        $new_mail->body = $novel_group->secret . " 이후로'" . $novel_group->title . "'이(가) 비밀 해제 됩니다.";
         $new_mail->from = Auth::user()->id;
         $new_mail->novel_group_id = $novel_group->id;
         $new_mail->save();
@@ -398,14 +520,15 @@ class NovelGroupController extends Controller
         }
     }
 
-    public function clone_for_publish($id)
+    public
+    function clone_for_publish($id)
     {
         try {
             $cloning_novel_group = NovelGroup::find($id);
             // novel_group to clone
             $new_novel_group = $cloning_novel_group->replicate();
             $new_novel_group->secret = Carbon::now();
-            $new_novel_group->title = $new_novel_group->title . "[15세 개정판]";
+            $new_novel_group->title = $new_novel_group->title . "[클린 버젼]";
             $new_novel_group->push();
             // novel_group cloned
 
@@ -418,10 +541,37 @@ class NovelGroupController extends Controller
                 $new_novel->novel_group_id = $new_novel_group->id;
                 $new_novel->push();
             }
-        } catch (Exception $e){
+        } catch (Exception $e) {
             abort(403, '처리 중 에러가 발생했습니다 관리자에게 문의하세요.');
         }
 
+    }
+
+    public function code_num_save(Request $request)
+    {
+        foreach ($request->all() as $item) {
+            $novel_group_code = NovelGroup::findOrFail($item['id']);
+            $novel_group_code->code_number = $item['code_number'];
+            $novel_group_code->save();
+        }
+
+        return response()->json($request);
+    }
+
+
+    public function recommend_order(Request $request)
+    {
+
+        foreach ($request->all() as $item) {
+            //999 used for sorting  asc in index method for admin
+            if($item['recommend_order'] ==null) $item['recommend_order']=999;
+
+            $novel_group_recommend_order = NovelGroup::findOrFail($item['id']);
+            $novel_group_recommend_order->recommend_order = $item['recommend_order'];
+            $novel_group_recommend_order->save();
+        }
+
+        return response()->json($request);
     }
